refactor: change auth state types
