+++ src/shared/api/supabase/types/auth-user.ts
@@ -19,16 +19,3 @@ export type {
   User,
   UserResponse
 }
-
-export type AuthEvent =
-  | 'SIGNED_IN'
-  | 'SIGNED_OUT'
-  | 'USER_UPDATED'
-  | 'TOKEN_REFRESHED'
-  | 'INITIAL_SESSION'
-  | 'UNKNOWN'
-
-export interface AuthState {
-  session: Session | null
-  event: AuthEvent
-}
