+++ src/shared/types/auth-identity.ts
@@ -1,7 +1,21 @@
-import type { Session, User } from '@shared/api'
+import type { AuthChangeEvent, Session, User } from '@supabase/supabase-js'
 
 export type AuthSession = Session
 export type AuthUser = User
+export type LowLevelAuthEvent = AuthChangeEvent
+
+export type AuthEvent =
+  | 'SIGNED_IN'
+  | 'SIGNED_OUT'
+  | 'USER_UPDATED'
+  | 'TOKEN_REFRESHED'
+  | 'INITIAL_SESSION'
+  | 'UNKNOWN'
+
+export interface AuthState {
+  session: AuthSession | null
+  event: AuthEvent
+}
 
 export type CurrentUserState = {
   user: AuthUser | null
