+++ src/shared/types/index.ts
@@ -1,2 +1,10 @@
+export type {
+  AuthEvent,
+  AuthSession,
+  AuthState,
+  AuthUser,
+  CurrentUserState,
+  LowLevelAuthEvent
+} from '../types/auth-identity'
 export type { AuthActionResult } from './register-result'
 export type { Theme, ThemePreference } from './theme-modes'
